+++ test/integration/contract_v2_test.exs
@@ -18,7 +18,6 @@ defmodule InfluxElixir.Integration.ContractV2Test do
     client: InfluxElixir.Client.HTTP,
     profile: :v2
 
-  alias InfluxElixir.Client.HTTP
   alias InfluxElixir.IntegrationHelper, as: H
 
   setup_all do